pyproject.toml
[tool.poetry]
name = "anverali-test"
version = "1.0.0"
description = ""
authors = ["Kirill Volychev <volychevk@gmail.com>"]
readme = "README.md"
packages = [{include = "anverali_test"}]

[tool.poetry.dependencies]
python = "^3.12"
fast-bitrix24 = "^1.7.1"
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
asyncache = "^0.3.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
flake8 = "^6.1.0"
flake8-black = "^0.3.6"
flake8-commas = "^2.1.0"
flake8-builtins = "^2.1.0"
flake8-variables-names = "^0.0.6"
pep8-naming = "^0.13.3"
reorder-python-imports = "^3.10.0"
autoflake = "^2.2.0"
setuptools = "^68.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
